fixed a bug with version

+++ src/main/kotlin/org/aydm/danak/service/FileService.kt
@@ -53,7 +53,7 @@ interface FileService {
      */
     fun delete(id: Long)
     fun saveAll(files: MutableList<FileDTO>):MutableList<FileDTO>
-    fun findAllLastVersion(version: Int):MutableList<FileDTO>
+    fun findAllLastVersion(version: Long):MutableList<FileDTO>
     fun findAllBelongsToVersion(version: Long): MutableList<FileDTO>
     fun findAllUpdates(v1: Long, v2: Long): MutableList<FileDTO>
 }
